refactor: utill 함수 파일명 변경 및 쓰지 않는 파일 제거
