+++ frontend/techpick/src/components/SideNavigationBar/CreateChildFolderInput.tsx
@@ -2,7 +2,7 @@
 
 import { useCreateFolder } from '@/queries/useCreateFolder';
 import type { FolderIdType } from '@/types/FolderIdType';
-import { isEmptyString } from '@/utils/string';
+import { isEmptyString } from '@/utils/isEmptyString';
 import { FolderPlus } from 'lucide-react';
 import {
   type KeyboardEvent,
