+++ frontend/techpick/src/utils/checkImageUrlByUrl.ts
@@ -1,6 +1,6 @@
-export function checkImageUrlByUrl(url: string | undefined): string {
+export const checkImageUrlByUrl = (url: string | undefined) => {
   if (!url || !url.trim().startsWith('http')) {
     return '/image/default_image.svg';
   }
   return url;
-}
+};
