+++ frontend/techpick/src/utils/isEmptyString.ts
