--- frontend/techpick/src/utils/portal.ts
@@ -1,12 +1,0 @@
-'ues client';
-
-import { PORTAL_CONTAINER_ID } from '@/constants/portalContainerId';
-
-export const getPortalContainer = () => {
-  const portalContainer = document.querySelector(`#${PORTAL_CONTAINER_ID}`);
-
-  if (!portalContainer) {
-    throw new Error('can not found portal container');
-  }
-  return portalContainer;
-};
